+++ src/commonMain/kotlin/com/inari/firefly/physics/movement/SimpleStepIntegrator.kt
@@ -7,7 +7,7 @@ import kotlin.math.abs
 
 class SimpleStepIntegrator  private constructor() : Integrator() {
 
-    override fun integrate(movement: EMovement, transform: ETransform, deltaTimeInSeconds: Long) {
+    override fun integrate(movement: EMovement, transform: ETransform, deltaTimeInSeconds: Float) {
         val velocity = movement.velocity
         if (movement.mass == 0f)
             return
@@ -20,7 +20,7 @@ class SimpleStepIntegrator  private constructor() : Integrator() {
         velocity.dy = velocity.dy + abs( (velocity.dy / movement.mass - 1f ) * 0.2f )
     }
 
-    override fun step(movement: EMovement, transform: ETransform, deltaTimeInSeconds: Long) =
+    override fun step(movement: EMovement, transform: ETransform, deltaTimeInSeconds: Float) =
         transform.move(movement.velocity.dx, movement.velocity.dy)
 
     companion object : SystemComponentSubType<Integrator, SimpleStepIntegrator>(Integrator, SimpleStepIntegrator::class) {
